Change in LiquidityTree multiplier calculation to uint256, added test

+++ contracts/LiquidityTree.sol
@@ -138,7 +138,9 @@ contract LiquidityTree {
         push(updatedNode, begin, end, leaf, ++updateId);
 
         // remove amount (percent of amount) from leaf to it's parents
-        withdrawAmount = (treeNode[leaf].amount * percent) / DECIMALS;
+        withdrawAmount = uint128(
+            (uint256(treeNode[leaf].amount) * percent) / DECIMALS
+        );
 
         updateUp(leaf, withdrawAmount, true, ++updateId);
 
@@ -324,7 +326,7 @@ contract LiquidityTree {
 
     /**
      * @dev remove amount from whole tree, starting from top node #1
-     * @param amount value to remove
+     * @param amount value to removeamount
      */
     function remove(uint128 amount) public {
         if (treeNode[1].amount >= amount) {
@@ -366,7 +368,7 @@ contract LiquidityTree {
         uint256 lAmount = treeNode[lChild].amount;
         uint256 sumAmounts = lAmount + treeNode[rChild].amount;
         if (sumAmounts == 0) return;
-        uint128 setLAmount = uint128((amount * lAmount) / sumAmounts);
+        uint128 setLAmount = uint128((uint256(amount) * lAmount) / sumAmounts);
 
         // update left and right child
         setAmount(lChild, setLAmount, updateId_);
@@ -467,8 +469,10 @@ contract LiquidityTree {
                     );
                 uint128 sumAmounts = lAmount + rAmount;
                 if (sumAmounts == 0) return;
-                uint128 forLeftAmount = (amount *
-                    ((lAmount * DECIMALS) / sumAmounts)) / DECIMALS;
+                uint128 forLeftAmount = uint128(
+                    (amount * ((uint256(lAmount) * DECIMALS) / sumAmounts)) /
+                        DECIMALS
+                );
 
                 // l in [begin,mid] - part in left child
                 pushLazy(
